+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `WriteOptions::attributes_indent`.
 - `Length::new_number`.
 - Text escaping before saving to file.
+- Enforced mutability. Many `Document`'s and `Node`'s methods require `&mut self` now.
+  This will prevent many runtime errors caused by borrowing `Rc` as mutable more than once.
 
 ### Changed
 - `writer` module is private now.
@@ -19,6 +21,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   input in debug mode now.
 - `postproc::resolve_stop_attributes` no longer converts `offset` attribute
   into a `Number` type, leaving it in a `Length` type.
+- This methods are require `&mut self` now:
+  - `Document`: create_element, create_node, append, drain.
+  - `Node`: detach, remove, drain, append, prepend, insert_after,
+    insert_before, text_mut, set_text, set_id, set_tag_name,
+    attributes_mut, set_attribute, set_attribute_checked, remove_attribute,
+    remove_attributes
+  - `postproc`: fix_rect_attributes, fix_poly_attributes
 
 ### Removed
 - `Node::attribute`. Use `node.attributes().get()` instead.
@@ -27,6 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `Node::set_link_attribute`. Use `Node::set_attribute` instead.
 - `Node::set_attribute_object`. Use `Node::set_attribute` instead.
 - All `AttributeValue::as_*` methods.
+- `Document::default`, because it was useless.
 
 ### Fixed
 - `postproc::resolve_stop_attributes` can be executed multiple times without errors now.